+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+* Now imessage can send attachment with `-a`.
+
 ## [v0.0.2](https://github.com/linjunpop/imessage/tree/v0.0.2)
 
 * Keep ruby 1.9 compatibility [@otzy007](https://github.com/otzy007) (https://github.com/linjunpop/imessage/pull/1)